Check if mLocationManager variable is initialized
In app first launch, after location permission is requested and location is disabled
mLocationManager is not initialized so app crash when is closed, with this workaround
that is fixed.

TODO: listen to location activation or request to user to activate it or find a better
method to handle location disabled after first launch.